A1/A_1/requirements.txt
aniso8601==9.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.30
protobuf==3.20.1
pytz==2022.4
PyYAML==6.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
